+++ pages/api/preview.ts
@@ -2,7 +2,7 @@ import {
   GITHUB_ACCESS_TOKEN_COOKIE_KEY,
   FORK_COOKIE_KEY,
   HEAD_BRANCH_COOKIE_KEY,
-} from '../../open-authoring/utils/cookieHelpers'
+} from '../../open-authoring/utils/repository'
 
 export default (req, res) => {
   const previewData = {
